Refactor, add button to vis window, add option to options menu
On refactoring, I've reorganized the source into a few packages to
help with organization. The gui package contains alloy4whole
things, the viz package contains alloy4viz things, and the server
package contains all server related code.

Added a button to the visualization window that opens the browser
page, as well as an option to the options menu that when enabled
will automatically open the browser page when Alloy starts up.