Add version info and enable next button from web page
Added both Sterling and Alloy version info to gradle build. The
Alloy version info is extracted from the Alloy JAR and placed
into the Sterling manifest. Added the sterling package and moved
the Sterling.java file into it, along with a new SterlingVersion
class that reads Sterling version information from the manifest.

Added "next" button functionality from the web page by modifying
the enumerator object in SimpleGUI. Now sets the subrunningTask
value dependent on whether or not the viz window is currently
visible. This allows us to maintain that same visibility state
when the computing of the next solution is finished. This way
the visualization window does not automatically open if it is
currently closed and the user requests the next solution from the
web.